messenger_auth_server/requirements.txt
APScheduler==3.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
